feat: history 기능
